chore(repo): master repository cleanup, documentation architecture, and stable uefi baseline
- Reorganize 600+ loose root files into structured docs/, tools/, tests/, and build/logs/ directories
- Restore concrete mojo IPC subsystem directory
- Author canonical documentation suite: The Book of ATOMS OS (Chapters 1-20), START_HERE.md, AI_ASSISTED_DEVELOPMENT.md, TESTING.md, KNOWN_ISSUES.md, MAINTENANCE.md
- Fix all links in README.md (107/107 valid links)
- Verify clean kernel build and pure UEFI QEMU boot

+++ kernel/vfs/vfs_legacy/include/vfs.h
@@ -58,7 +58,7 @@ int vfs_open(const char* path);
 int vfs_read(int fd, void* buffer, uint32_t size);
 int vfs_write(int fd, void* buffer, uint32_t size);
 int vfs_pread(int fd, void* buffer, uint32_t size, uint64_t offset);
-int vfs_seek(int fd, uint64_t offset, int whence);
+int64_t vfs_seek(int fd, int64_t offset, int whence);
 int vfs_close(int fd);
 int vfs_readdir(const char* path, int index, vfs_dirent_t* out_entry);
 int vfs_mkdir(const char* path);
